rename all components to temporary suffix
