+++ src/components/settingsComp.js
@@ -3,7 +3,7 @@ const React = require('react'),
 	validator = require('validator'),
 	Link = require('react-router').Link,
 	bindActionCreators = require('redux').bindActionCreators,
-	Header = require('./../containers/header'),
+	Header = require('./../containers/headerComp'),
 	actions = require("../actions"),
 	PasswordDialog = require('./../containers/password-dialog');
