+++ src/containers/coverComp.js
