+++ src/containers/wrapperComp.js
@@ -3,7 +3,7 @@ var React = require('react'),
   {withRouter} = require('react-router'),
   cookie = require('./../../lib/cookie'),
   actions = require('../actions'),
-  LastUserSelector = require('../components/lastUserSelector'),
+  LastUserSelector = require('../components/lastUserSelectorComp'),
   Login = require('./../auth/Login');
 
 var Wrapper = React.createClass({
