+++ src/dashboard/dashboardComp.js
@@ -1,7 +1,7 @@
 const React = require('react'),
 	ReactRedux = require('react-redux'),
-	EditImageHeader = require('./../containers/cover'),
-	Header = require('./../containers/header'),
+	EditImageHeader = require('./../containers/coverComp'),
+	Header = require('./../containers/headerComp'),
 	actions = require('../actions'),
 	moment = require('moment'),
 	_ = require('lodash'),
